requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
APScheduler==3.11.2
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
cryptography==46.0.6
debugpy==1.8.20
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
executing==2.2.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
griffelib==2.0.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
narwhals==2.19.0
nest-asyncio==1.6.0
numpy==2.4.4
openai==2.30.0
openai-agents==0.13.2
packaging==26.0
pandas==3.0.2
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.4
postgrest==2.28.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
python-multipart==0.0.22
pyzmq==27.1.0
realtime==2.28.3
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
sendgrid==6.12.5
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sse-starlette==3.3.4
stack-data==0.6.3
starlette==1.0.0
storage3==2.28.3
streamlit==1.56.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
types-requests==2.33.0.20260327
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.8
yarl==1.23.0
zstandard==0.25.0
